pyproject.toml
[project]
name = "slurm-email-notifs"
version = "0.1.0"
description = "Email notifications for SLURM jobs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
]

[project.scripts]
slurm-notify = "slurm_email_notifs.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
